Goatpeople/inventory (#6)
* Added basic item and inventory system

This definitely needs to be fleshed out, and main should include imports, but I'm just saving the past week of progress. Been working on the local branch for too long without committing.

* QUESTIONS - Asking questions about item dicts and interactions

I modified the 'item.py' file quite heavily in order to make answering questions easier.

This is a tutorial commit, so no need to get fancy.